Improvements for search component
